Fix build.
MFC after:	2 weeks
X-MFC-With:	r345900


git-svn-id: svn+ssh://svn.freebsd.org/base/head@345901 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sbin/fsck_msdosfs/dir.c
@@ -488,7 +488,7 @@ check_subdirectory(int f, struct bootblock *boot, struct dosDirEntry *dir)
 
 	off *= boot->bpbBytesPerSec;
 	if (lseek(f, off, SEEK_SET) != off ||
-	    read(f, buf, boot->bpbBytesPerSec) != boot->bpbBytesPerSec) {
+	    read(f, buf, boot->bpbBytesPerSec) != (ssize_t)boot->bpbBytesPerSec) {
 		perr("Unable to read directory");
 		free(buf);
 		return FSFATAL;
